refactor: remove waterfall Traces tab, rename Summaries to Traces
Removes the raw OTLP waterfall view (Traces tab) and renames the
human-readable session timeline (Summaries) to Traces in its place.
Cleans up associated state signals, CSS, and updates Help and README
to reflect the new tab name and removed waterfall references.
