+++ media/dashboard.css
@@ -511,31 +511,6 @@ body {
   overflow-y: auto;
   max-height: 70vh;
 }
-.wf-controls {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  margin-bottom: 10px;
-  font-size: 12px;
-}
-.wf-controls label {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  cursor: pointer;
-}
-.wf-controls button {
-  background: var(--accent);
-  color: var(--bg);
-  border: none;
-  border-radius: 4px;
-  padding: 4px 12px;
-  font-size: 11px;
-  cursor: pointer;
-}
-.wf-controls button:hover {
-  opacity: 0.85;
-}
 .wf-trace-group {
   margin-bottom: 16px;
 }
@@ -619,7 +594,6 @@ body {
   width: 16px;
   flex-shrink: 0;
 }
-.wf-chevron,
 .sw-chevron {
   font-size: 8px;
   width: 12px;
@@ -663,36 +637,6 @@ body {
   height: 100%;
   border-radius: 3px;
 }
-.wf-bar .wf-dur {
-  position: absolute;
-  right: -4px;
-  top: 50%;
-  transform: translate(100%, -50%);
-  font-size: 9px;
-  color: var(--muted);
-  white-space: nowrap;
-  padding-left: 4px;
-}
-.wf-bar .wf-tip {
-  display: none;
-  position: absolute;
-  bottom: 22px;
-  left: 0;
-  background: var(--card-bg);
-  border: 1px solid var(--border);
-  border-radius: 4px;
-  padding: 8px 10px;
-  font-size: 10px;
-  white-space: pre-line;
-  z-index: 10;
-  pointer-events: none;
-  min-width: 250px;
-  max-width: 400px;
-  line-height: 1.5;
-}
-.wf-bar:hover .wf-tip {
-  display: block;
-}
 .wf-info {
   width: 90px;
   min-width: 90px;
@@ -700,17 +644,6 @@ body {
   padding-right: 8px;
   font-size: 10px;
   color: var(--muted);
-}
-.wf-detail {
-  display: none;
-  padding: 8px 12px 8px 44px;
-  background: var(--hover);
-  font-size: 11px;
-  border-bottom: 1px solid var(--border);
-  line-height: 1.7;
-}
-.wf-detail.open {
-  display: block;
 }
 .wf-detail-row {
   display: flex;
@@ -723,25 +656,6 @@ body {
 }
 .wf-detail-val {
   word-break: break-all;
-}
-.wf-prompt-block {
-  padding: 8px 12px 8px 12px;
-  background: var(--hover);
-  font-size: 12px;
-  border-bottom: 1px solid var(--border);
-  line-height: 1.6;
-  white-space: pre-wrap;
-  word-break: break-word;
-}
-.wf-prompt-label {
-  font-weight: 600;
-  color: var(--accent);
-  font-size: 10px;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-}
-.wf-prompt-text {
-  color: var(--fg);
 }
 
 /* media/src/styles/summaries.css */
