+++ media/src/tabs/Help.tsx
@@ -10,8 +10,7 @@ const VIEWS: [string, string][] = [
   ['Automation',      'Automated prompts triggered when session thresholds are crossed. Configure per-agent automations for Loop Breaker, Turn Limit Wrap-up, and Context Dump. In the VS Code extension, automations show a notification or open the agent chat directly; in standalone mode they write to a file-based relay.'],
   ['Tokens',          'Token consumption aggregated by span name and per session, sorted from highest to lowest.'],
   ['Latency',         'Span durations as a color-coded grid, helping identify which operations are consistently slow.'],
-  ['Summaries',       'A human-readable session waterfall with LLM decisions, tool arguments and results, token usage per step, and background overhead breakdown.'],
-  ['Traces',          'Raw OTLP spans as horizontal bars on a time axis, preserving the full parent-child nesting hierarchy and exact timing.'],
+  ['Traces',          'A human-readable timeline of each session — LLM calls with decisions, tool calls with arguments and results, token usage per step, and background overhead breakdown.'],
   ['Files',           'Files created or modified by the agent, organized by session with inline before/after diffs showing exactly what changed.'],
   ['Flow',            'LLM turns and tool calls visualized as a semantic graph — one node per turn, one per unique tool, edges weighted by call frequency. Supports zoom, pan, and playback animation.'],
   ['Agents',          'Side-by-side comparison of Copilot, Claude, and Codex with per-agent token totals, cache rates, time-to-first-token, and top tools, plus a full session history table.'],
@@ -42,7 +41,6 @@ const TERMS: [string, string][] = [
   ['Output Ratio',           'Percentage of total tokens that are output (generated by the model). In cached agentic coding sessions this can be naturally tiny, so AgentLens no longer uses it as a standalone alert.'],
   ['Prompt',                 'The text you type into the AI chat to request work. Each prompt initiates a new session.'],
   ['Request',                'The user-visible message sent to the agent in a single prompt. In OTEL terms, the request anchor differs by agent: Copilot uses invoke_agent, Claude uses claude_code.interaction, and Codex is normalized from prompt log events.'],
-  ['Retain Spans',           "A waterfall checkbox option. When checked, spans from previous prompts are kept visible alongside new ones."],
   ['Session',                'A single prompt-to-response cycle. Starts when you send a prompt and ends when the agent delivers its final response. AgentLens normalizes different Copilot, Claude, and Codex OTEL shapes into this shared model.'],
   ['Span',                   'A single timed operation recorded by OpenTelemetry. AgentLens displays true trace spans and normalized log events with a span-like name, duration, and attributes.'],
   ['Span ID',                'A unique identifier for a single span within a trace. Used to establish parent-child relationships between operations.'],
@@ -54,7 +52,7 @@ const TERMS: [string, string][] = [
   ['Trace ID',               'A unique identifier linking all spans belonging to the same session/request.'],
   ['Turn',                   'One LLM call within a session. A multi-turn session involves the agent calling the LLM, executing tools, then calling the LLM again.'],
   ['TTFT',                   'Time to First Token — the latency between sending a prompt and receiving the first token of the response.'],
-  ['Waterfall',              'A visualization where spans are displayed as horizontal bars on a time axis, with nesting depth shown by indentation.'],
+  ['Waterfall',              'A span visualization where operations are displayed as horizontal bars on a time axis, with nesting depth shown by indentation. Used in OpenTelemetry tooling; AgentLens surfaces span timing data through the Latency tab instead.'],
 ]
 
 function termId(term: string): string {
@@ -116,7 +114,7 @@ const AGENT_OTEL_SHAPES: Array<{
     agent: 'Codex',
     format: 'Primarily flat OTLP log records (structured JSON events sent to /v1/logs), not trace spans. Each session is a stream of log events grouped by conversation and turn identifiers. Adding trace_exporter to ~/.codex/config.toml also emits timing spans to /v1/traces. Both the CLI and the VS Code extension read the same config file.',
     coverage: 'With the recommended configuration (log_user_prompt = true and both exporters set): prompt text, token counts, model name, TTFT, tool names, tool arguments, tool results, and span timing are all present.',
-    gaps: 'Trace and Timeline tabs have less span granularity than Copilot or Claude Code since Codex is primarily log-based. Without trace_exporter, span waterfall data is limited.',
+    gaps: 'The Traces timeline has less span granularity than Copilot or Claude Code since Codex is primarily log-based. Without trace_exporter, timing data is limited.',
   },
 ]
 
@@ -383,7 +381,7 @@ function AgentOtelSection() {
             </div>
           ))}
         </div>
-        <p style="margin-top:14px;font-size:12px;color:var(--muted)">The practical effect: <a href="#gl-trace">Traces</a> and Timeline stay closest to the raw OTEL structure, while Efficiency, Summaries, Recommendations, Alerts, Automation, Agents, and Flow use the normalized session model so the three agents can be compared side by side.</p>
+        <p style="margin-top:14px;font-size:12px;color:var(--muted)">The practical effect: Traces and Timeline stay closest to the raw OTEL structure, while Efficiency, Recommendations, Alerts, Automation, Agents, and Flow use the normalized session model so the three agents can be compared side by side.</p>
       </div>
     </div>
   )
@@ -476,7 +474,7 @@ function LoopsSection() {
           />
           <LoopBlock id="help-context-accumulation" title="Infinite Loop — Context Accumulation"            why={`<a href="#gl-input-tokens">Input tokens</a> grew by 30,000+ across 4+ calls while <a href="#gl-output-ratio">output-to-input ratio</a> collapsed by 70%+. The agent is consuming context while producing less output.`}
             example="First call: 8K in → 600 out (7.5%). Last call: 65K in → 80 out (0.12%). Five turns reading the same files without edits."
-            steps={`<li>Stop immediately — cost compounds with no progress.</li><li>Start fresh with a focused prompt stating what was already read.</li><li>Include the specific target state, not just the problem.</li><li>Use the Summaries tab to review what was accomplished.</li>`}
+            steps={`<li>Stop immediately — cost compounds with no progress.</li><li>Start fresh with a focused prompt stating what was already read.</li><li>Include the specific target state, not just the problem.</li><li>Use the Traces tab to review what was accomplished.</li>`}
             impact="Catching at 4 calls instead of 10 saves ~390,000 input tokens at peak context size."
           />
         </div>
